% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_region_split_group}
\alias{add_region_split_group}
\title{Add Split Group to Region Data}
\usage{
add_region_split_group(data, split_lims, type = c("orf", "utr"))
}
\arguments{
\item{data}{Data frame describing ORFs or UTRs.}

\item{split_lims}{Split limits data frame, result of \code{get_split_lims()}.}

\item{type}{Type of region data. Either "orf" or "utr".}
}
\value{
Data frame with additional grouping variable(s).
}
\description{
Adds position split \code{group} variable to data frames when using facets
in \code{plot_summary_map()} function.
}
