% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_split_group}
\alias{add_split_group}
\title{Add Split Group to Data}
\usage{
add_split_group(data, nt_num, num_facets)
}
\arguments{
\item{data}{Data frame with \code{position} variable.}

\item{nt_num}{Number of nucleotides of the sequence.}

\item{num_facets}{Number of facets or splits.}
}
\value{
Data frame with additional \code{group} variable.
}
\description{
Adds position split \code{group} variable to data frames when using facets
in \code{plot_summary_map()} function.
}
