% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pairs.R
\name{aggregate_base_pairs_prob}
\alias{aggregate_base_pairs_prob}
\title{Aggregate Base-Pair Probabilities}
\usage{
aggregate_base_pairs_prob(windowed_folds, freq_cutoff = 0.5, prob_cutoff = 0.9)
}
\arguments{
\item{windowed_folds}{List of \code{RNAfold} results for each sliding window. Result of \code{fold()} function.}

\item{freq_cutoff}{Frequency cut-off.}

\item{prob_cutoff}{Probability cut-off.}
}
\value{
Data frame specifying folding probability of the base-pair \code{(pos_i, pos_j)}.
}
\description{
Aggregate Base-Pair Probabilities
}
