% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arc_segment}
\alias{arc_segment}
\title{Define Arc Segment}
\usage{
arc_segment(c, r, angles = c(0, pi), length = 100)
}
\arguments{
\item{c}{Vector c(x, y) defining center of the arc.}

\item{r}{Radius of the arc.}

\item{angles}{Angles between which the arc segment is defined.}

\item{length}{Length of points defining the arc segment.}
}
\value{
Data frame of \code{(x, y)} points defining the arc segment.
}
\description{
Define Arc Segment
}
