% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{build_UTRs_data}
\alias{build_UTRs_data}
\title{Build UTRs Data Frame}
\usage{
build_UTRs_data(utr5_lims = NULL, utr3_lims = NULL)
}
\arguments{
\item{utr5_lims}{Vector of lower and upper limits of UTR 5'.}

\item{utr3_lims}{Vector of lower and upper limits of UTR 3'}
}
\value{
Data frame with \code{start} and \code{end} of each UTR.
}
\description{
Build UTRs Data Frame
}
