% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pairs.R
\name{clean_base_pairs}
\alias{clean_base_pairs}
\title{Clean Base-Pairs Data}
\usage{
clean_base_pairs(base_pairs_data)
}
\arguments{
\item{base_pairs_data}{Data frame specifying folding probability of the base-pair \code{(pos_i, pos_j)}.}
}
\value{
Data frame with unique base-pair foldings.
}
\description{
Clean base-pairs data to ensure no nucleotide is paired with more than one nucleotide.
}
