% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\alias{calc_shannon_entropy}
\alias{aggregate_shannon_entropy}
\title{Calculate and Aggregate Shannon's Entropy}
\usage{
calc_shannon_entropy(dotplot_data, prob_cutoff = 0)

aggregate_shannon_entropy(windowed_folds, prob_cutoff = 0)
}
\arguments{
\item{dotplot_data}{Dot-plot data frame.}

\item{prob_cutoff}{Probability cut-off.}

\item{windowed_folds}{List of \code{RNAfold} results for each sliding window. Result of \code{fold()} function.}
}
\description{
Calculate and Aggregate Shannon's Entropy
}
