% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Fold with Sliding Windows}
\usage{
fold(
  filename,
  winsize,
  stepsize,
  same_num_samples = TRUE,
  rnafold_params = "-p",
  verbose = FALSE
)
}
\arguments{
\item{filename}{The name of the file which the sequences in FASTA format are to be read from.}

\item{winsize}{Window size of the sliding windows.}

\item{stepsize}{Stepsize of the sliding windows.}

\item{same_num_samples}{If TRUE, fold will perform additional foldings at the beginning and end of the sequence. This allows the beginning and end of the sequence to have the same amount of samples/windows go through it. This is specially important for the calculation of Shannon's entropy.}

\item{rnafold_params}{Parameters used by RNAfold. The default is '-p'.}

\item{verbose}{If TRUE, fold will print information of performance}
}
\value{
List of RNAfold results for each sliding window
}
\description{
Fold with Sliding Windows
}
