% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_arc_trajectory}
\alias{get_arc_trajectory}
\title{Calculate Arc Trajectory Between Two Points}
\usage{
get_arc_trajectory(i, j)
}
\arguments{
\item{i}{Position of tarting point.}

\item{j}{Position of ending point.}
}
\value{
Arc trajectory.
}
\description{
Calculates the arc trajectory between two points on the \code{x} axis.
}
