% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{get_ORFs}
\alias{get_ORFs}
\title{Get Open Reading Frame Data}
\usage{
get_ORFs(seq, calc_groups = TRUE)
}
\arguments{
\item{seq}{Sequence input.}

\item{calc_groups}{If TRUE, the ORFs will be separated into non-overlapping groups.}
}
\value{
Data frame specifying \code{start}, \code{end}, and optionally \code{interval_group} for each ORF.
}
\description{
Find open reading frames (ORFs) of the sequence. Optionally, the ORFs can
be separated into non-overlapping groups, allowing to easily plot them as
genome map plots.
}
