% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_split_intervals}
\alias{get_split_intervals}
\title{Get Position Split Intervals Data}
\usage{
get_split_intervals(nt_num, num_facets)
}
\arguments{
\item{nt_num}{Number of nucleotides of the sequence.}

\item{num_facets}{Number of facets or splits.}
}
\value{
Split intervals data frame.
}
\description{
Get position split intervals data frame when using facets in
\code{plot_summary_map()} function.
}
