% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_split_lims}
\alias{get_split_lims}
\title{Get Position Split Limits Data}
\usage{
get_split_lims(nt_num, num_facets)
}
\arguments{
\item{nt_num}{Number of nucleotides of the sequence.}

\item{num_facets}{Number of facets or splits.}
}
\value{
Split limits data frame.
}
\description{
Get position split limits data frame when using facets in
\code{plot_summary_map()} function.
}
