% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_structure_map}
\alias{plot_structure_map}
\title{Plot Structure Map from Sliding Windows}
\usage{
plot_structure_map(
  windowed_folds,
  utr5_lims = NULL,
  utr3_lims = NULL,
  num_facets = 3,
  freq_cutoff = 0.5,
  prob_cutoff = 0.9,
  plot_list = c("entropy", "bpp", "orf", "utr")
)
}
\arguments{
\item{windowed_folds}{List of \code{RNAfold} results for each sliding window. Result of \code{fold()} function.}

\item{utr5_lims}{Vector of lower and upper limits of UTR 5'.}

\item{utr3_lims}{Vector of lower and upper limits of UTR 3'.}

\item{num_facets}{Number of facets or splits.}

\item{freq_cutoff}{Frequency cut-off.}

\item{prob_cutoff}{Probability cut-off.}

\item{plot_list}{List of plots to display. Options are \code{c("entropy", "bpp", "orf", "utr")}, the order of which will be respected.}
}
\value{
A \code{patchwork} object.
}
\description{
Plot structure map from sliding windows folding results.
}
