% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\alias{plot_entropy}
\alias{plot_base_pairs}
\alias{plot_orfs}
\alias{plot_utrs}
\title{Individual Plots}
\usage{
plot_entropy(data, split_lims, hide_x = FALSE)

plot_base_pairs(data, prob_cutoff = 0.9, split_lims, hide_x = FALSE)

plot_orfs(data, split_lims, hide_x = FALSE)

plot_utrs(data, split_lims, hide_x = FALSE)
}
\arguments{
\item{data}{Data frame for entropy, base-pair probabilities, ORFs, or UTRs.}

\item{split_lims}{Split limits data frame, result of \code{get_split_lims()}.}

\item{hide_x}{If TRUE, the \code{x} axis will not be shown.}

\item{prob_cutoff}{Probability cut-off.}
}
\value{
A \code{ggplot} object.
}
\description{
Individual Plots
}
