% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ps_dotplot}
\alias{read_ps_dotplot}
\title{Read PostScript Dot-Plot File}
\usage{
read_ps_dotplot(filename)
}
\arguments{
\item{filename}{The name of the PostScript file which dot-plot data is to be read from.}
}
\value{
Dot-plot data frame.
}
\description{
Read PostScript Dot-Plot File
}
