% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Fold with Sliding Windows}
\usage{
fold(
  filename,
  winsize,
  stepsize,
  same_num_samples = TRUE,
  rnafold_params = "-p",
  verbose = FALSE
)
}
\arguments{
\item{filename}{The name of the file which the sequences in FASTA format are to be read from.}

\item{winsize}{Window size of the sliding windows.}

\item{stepsize}{Step size of the sliding windows.}

\item{same_num_samples}{If TRUE, fold will perform additional foldings at the beginning and end of the sequence. This allows the beginning and end of the sequence to have the same amount of samples/windows go through it. This is specially important for calculating Shannon's entropy.}

\item{rnafold_params}{Parameters used by \code{RNAfold}. The default is \code{'-p'}.}

\item{verbose}{If TRUE, fold will print information of performance.}
}
\value{
List of \code{RNAfold} results for each sliding window.
}
\description{
Perform folding inference using sliding windows and \code{RNAfold} from the
ViennaRNA Package 2.0 by Lorenz et al. \url{doi:10.1186/1748-7188-6-26}.
}
